requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Flask-Cors
python-dotenv
requests
razorpay
gunicorn
pymongo[srv]
Werkzeug
setuptools
